feat: isolate filterThProductByWorkerCodeAndDateBetween() in a new service/test
